### 2021-12-17 #4003 @whedon
[PRE REVIEW]: ReMove: A Configuration Tool for Robot-Assisted Experiments in Rehabilitation
**Submitting author:** @Rehablab (<a href="http://orcid.org/0000-0001-8781-301X">Qiang Xu</a>)
**Repository:** <a href="https://github.com/Rehablab/ReMove" target ="_blank">https://github.com/Rehablab/ReMove</a>
**Version:** v1.0.0
