### 2021-12-17 #4003 @kyleniemeyer
Hello @Rehablab, the JOSS editorial board is going to review your submission to see if it meets our requirements; we will get back to you within a week or two on this.
