### 2021-12-18 #4003 @danielskatz
@rehablab - I suspect this doesn't compile because of one or more of three things:

* the blank orcid line should be remove - if there is no orcid, this line isn't needed

* `    - name: Chuanxin M. Niu^[corresponding author&#58;&nbsp;minos.niu&#64;gmail.com]`

which you could try changing to just to
`    - name: Chuanxin M. Niu^[corresponding author]`

* or the blank line between the affiliations and the date, which you could remove.

In addition, you could work on the possibly missing DOIs that whedon suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@whedon check references` to check again, and the command `@whedon generate pdf` when the references are right (or you've made changes in the .md file) to make a new PDF. Whedon commands need to be the first entry in a new comment.
