### 2021-12-18 #4003 @danielskatz
Also, looking at your paper, you have put the references in manually rather than tagging them so they are built - see the example paper: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography
