### 2021-12-20 793bdb9
:memo:remove the references in paper.md
