+++ paper/paper.md
@@ -22,8 +22,7 @@ authors:
       orcid:
     - name: Teck Hong Koh
       affiliation: "3"
-      orcid:
-    - name: Chuanxin M. Niu^[corresponding author&#58;&nbsp;minos.niu&#64;gmail.com]
+    - name: Chuanxin M. Niu^[corresponding author&#58; minos.niu&#64;gmail.com]
       affiliation: "1, 4"
       orcid: 0000-0001-6013-2492
 affiliations:
