### 2021-12-20 c7c2d86
:memo:update the config of figures
