### 2022-01-10 #4003 @adi3
@federeghe @sneakers-the-rat @vmatos @AustinTSchaffer @destogl @ekorudiawan - Would any of you be interested in reviewing this paper on robotics? We would greatly appreciate your help!
