### 2022-01-17 #4071 @whedon
[REVIEW]: ReMove: A Configuration Tool for Robot-Assisted Experiments in Rehabilitation
**Submitting author:** @Rehablab (<a href="http://orcid.org/0000-0001-8781-301X">Qiang Xu</a>)
**Repository:** https://github.com/Rehablab/ReMove
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0

> Not sure to whom the @rehablab handle belongs to, but all but one commit are signed by author with email that matches first author's name

> I was able to figure out how to get Unity installed, but do not know if I installed the software correctly because there is no documentation to speak of.

> I was unable to get any of the scenes to run. None of the buttons seem to be connected to actions, and in the TrialScene i was unable to tell how to load a script. Please let me know how it is possible for me to run a task for testing -- I see that it is possible to use a mouse and would be happy to do that. I was also unable to get the tests to run using the instructions (Window > General > Test Runner) -- all that happened was the screen displayed the text "Display 1 No Cameras rendering"

> I would like to see more information about the state of the field here, what other products like this haptic controller exist, what the other interfaces are like, even just references to them so that I can gauge what I should expect the software to be capable of. I'm in general pretty lax on this checkbox so i'm filling it but let's call that a "nice to have" rather than "need to have" thing.

- [ ] **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution

> There aren't really any instructions so to speak, I had to figure out how Unity works, download an old version from their archives, and just sort of feel my way through it. This needs to be clearer!

> There is an example of a trial script, but I was unable to figure out how to load it and use it in the GUI, which I also could not figure out.

> There is no API documentation that I could find. I would like to see at least some general description of the modules and how I would understand the system, especially if it is intended to be expanded to other devices in the future -- as is I can't tell what that would take after reading the device drivers for ~10m or so.

> There are tests! I was unable to get them to run, and nothing was showing in the console. Would love to have the author's guidance on this.

> Did not see any, though if the usual caveat "open an issue" applies that might work, but would like to see contribution guidelines as part of the API documentation

> I would like to have a lot more here -- what is unique about this hardware system, and what are other systems already capable of doing? What are the general types of parameterizations you would want a system like this to do? eg. I am not sure what a "type of resistance" means

> There are in-text citations but they are not included in the references list
