### 2022-01-25 Rehablab/ReMove#1 @AustinTSchaffer
JOSS Review: Installation Instructions
Hello @Rehablab. I'm working through the JOSS review for this project. The review guidelines state that the submission should have the following 2 criteria:

> Installation: Does installation proceed as outlined in the documentation?
> Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

Thank you for providing the dependencies required to install and run this software. However, could you please add detailed installation instructions, or a link to Unity documentation on how to install a package such as this? Downstream users that are not Unity experts (including myself) might have issues figuring out how to install this project.

Related to: https://github.com/openjournals/joss-reviews/issues/4071
