### 2022-01-25 Rehablab/ReMove#2 @AustinTSchaffer
JOSS Review: Community guidelines
Hello again. Could you please add some information to the project's README with regards to:

>  Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Seeing that this project has an MIT license, you could satisfy this point by simply clarifying whether you're open to issue and pull request submissions on this repository.

Related to: https://github.com/openjournals/joss-reviews/issues/4071
