### 2022-01-25 Rehablab/ReMove#1 @AustinTSchaffer
Further, do you have any resources on how this software communicates with the M2PRO in order to adjust its parameters? I don't have any experience with M2PROs or any other similar hardware platforms, so I can't tell how simple or complicated of a process it would be to properly configure the parameters of the device with this software.
