### 2022-01-25 Rehablab/ReMove#3 @AustinTSchaffer
JOSS Review: Documentation Issues
I opened a 3rd issue for this, since the comments are a little less concrete than what's described in Issues #1 and #2.

> A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

This project's README contains some information in its Overview and About sections that cover what the project does, but it's not clear what exact problem the project solves until you've read the entire readme and the accompanying JOSS paper. The statement of need could simply use some extrapolation in this project's README by expanding on what's currently written, or lifting some of the information from your JOSS paper.

> Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

There are some pictures/graphics in your paper that you could/should also reference in the project's README as to what a user could/would/should expect when the run/use the software.

> Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

This section probably needs the most work. Could you please expand on how researchers might edit the CSV configuration file to achieve the desired outcome? What is a good starting point for the CSV configuration? What are the units of the mass and damping values in the configuration file?

Related to: https://github.com/openjournals/joss-reviews/issues/4071
