### 2022-01-25 #4071 @AustinTSchaffer
@adi3 I have some concerns about this submission that I believe extend beyond minor technicalities. In my view, this project misses the mark on [the JOSS guidelines for substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort), apart from the total number of lines of code. I am not sure how to proceed with the review of this project, so I'd just like to summarize my thoughts here:

> JOSS requires that software should be feature-complete [...] and designed for maintainable extension (not one-off modifications of existing tools).

The project's README makes mention of 2 "upgrades and new features" which I believe are ideals of what this software could achieve, but are not actually implemented in the source code.

The paper also mentions that this software can be used to adjust parameters of an "M2PRO, a 2D haptic robot", then mentions that "the support will expand to other available makes and models including Phantom series, HapticMASTER, etc." I believe that this claim could lead someone to believe that the program can be easily extended to work with 2D haptic robots from any vendor. I also believe that this is an indication that the project is a one-off modification of the M2PRO and is not a general-use tool.

> JOSS requires that software should be [...] packaged appropriately according to common community standards for the programming language being used

From what I can tell, this project is not packaged for distribution in a meaningful way. There are also little to no installation and usage instructions. Based on the wording of the project's dependencies, it seems to me that the project requires users to run the project from an IDE. This is not necessarily a deal-breaker for me, but in combination with the other points, I believe it further reduces the usability and applicability of the software to other research projects.

> [...] software should be a significant contribution to the available open source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler).

The paper's statement of need states that there is no existing software that allows researchers to adjust parameters of 2D haptic robots outside of editing source code. If this software project allowed researchers to tweak the parameters as part of the program's User Interface, I think that would be an excellent improvement. Unfortunately, I don't believe that having researchers edit a CSV file is significantly more intuitive than having researchers edit parameters in some other program's source code. This is of course an entirely subjective opinion, so I am willing to concede on this point if the authors add support to the UI for adjusting these settings, or add more documentation that would aid another researcher in adjusting the program's input CSV file.
