### 2022-02-01 #4071 @sneakers-the-rat
I feel similarly to @AustinTSchaffer - I need more information to complete my review, I've tried to list this in my checklist at the top. I was able to get a version of Unity that was compatible with the project, but once I opened it I couldn't get anything to work, outside of the trial panel none of the buttons seemed connected to functions, and the tests wouldn't run (said "no cameras rendering").

At a minimum I'll need some additional installation and API documentation to get me to a point where I can run the program and assess functionality, I'm fine with just testing with my mouse, or even seeing some of the automated tests run, but at the moment I can't finish my review.
