### 2022-02-15 #4071 @Rehablab
We thank the reviewers for the thoughtful comments. Since the questions were mostly about the relevance of ReMove to a wider range of devices and audience, this would require a reflection on our strategy and roadmap. However, we believe that ReMove still caters for more than just ourselves. We would like a 2-3 weeks for the revision of our manuscript.
